+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * Add storage counts to storage inventory screens
 * Make durability repair of lazurite tools with an ingot be based off of original ingredients rather than automatically 100%
-* Add power generation indiciator to Stirling Engine
+* Add power generation indicator to Stirling Engine
+* Add power draw indicator to Network Interface
 
 ## 26.1.2.2